pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
    name = "calypsokit"
    dynamic = ["version"]
    description = ""
    authors = [
        {name = "Xiaoshan Luo", email = "luoxs@calypso.cn"},
        {name = "Zhenyu Wang", email = "wzy@calypso.cn"},
    ]
    dependencies = [
        "numpy",
        "pandas",
        "pymongo",
        "python-dotenv",
    ]
    requires-python = ">=3.7.4"
    readme = "README.md"
    license = {text = ""}
    maintainers = [
    ]
    keywords = []
    classifiers = [
        "Programming Language :: Python :: 3.11",
        "Operating System :: POSIX :: Linux",
    ]

[project.optional-dependencies]
    docs = [
        "sphinx>=5.3.0",
        "myst-parser>=1.0.0",
        "sphinx-rtd-theme>=1.2.0",
        "numpydoc>=1.5.0",
    ]
    test = [
        "coverage"
    ]

[project.urls]
    # homepage = "***"
    # repository = ""
    # documentation = ""

[project.scripts]

[tool.setuptools]
    packages = ["calypsokit"]

[tool.setuptools_scm]
    write_to = "calypsokit/_version.py"

[tool.black]
    skip-string-normalization = true
    max-length = 88

[tool.isort]
    profile = "black"

[tool.mypy]
    follow-imports = "silent"
    show-column-numbers = true